+++ include/NeoFOAM/compatibility/fvSolution.hpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 //
-// SPDX-FileCopyrightText: 2023 FoamAdapter authors
-/* This file implements comparison operator to compare OpenFOAM and corresponding FoamAdapter fields
+// SPDX-FileCopyrightText: 2023 NeoFOAM authors
+/* This file implements comparison operator to compare OpenFOAM and corresponding NeoFOAM fields
  * TODO the comparison operator only make sense for testing purposes
  * so this should be part of the tests
  */
@@ -10,7 +10,7 @@
 #include "NeoN/core/dictionary.hpp"
 
 
-namespace FoamAdapter
+namespace NeoFOAM
 {
 
 void updateSolver(NeoN::Dictionary& solverDict);
@@ -19,4 +19,4 @@ void updatePreconditioner(NeoN::Dictionary& solverDict);
 
 NeoN::Dictionary mapFvSolution(const NeoN::Dictionary& solverDict);
 
-} // namespace FoamAdapter
+} // namespace NeoFOAM